meta-agl-bsp: Add overrides to rcar3 BSP changes
The various configuration changes done by the include files for the
rcar3 machines in meta-agl-bsp prove problematic when using a rcar3
machine in a multiconfig configuration with different machines in
other configs.  This can potentially be avoided in the future with
a significant rework of the local.conf generation in aglsetup.sh,
but for now, too avoid issues, add rcar-gen3 or h3ulcb overrides to
the variable changes in those files.  That ensures that the changes
will only apply to the specific machine being targeted.

So far this has only been tested with the h3ulcb, it is likely that
more fixes are required for other rcar3 machines, as well as the
machines from other BSPs.

Bug-AGL: SPEC-4618

Change-Id: Ifa6735337e673a92c56d5457b97d09a5b86cc0e1
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28185
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
(cherry picked from commit 3274485826c6c1c719cdcff03ce3dbd2d8120ce8)
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28277