weston/0001-simple-touch-Add-maximized-fullscreen: Add maximized/fullscreen for simple-touch
simple-touch wasn't able to resize itself to max/fullscreen to this
patch adds support for that. It is quite useful to have this client
test out touch support, so we can re-use it whenever, for instance in
the virtio aarch64 machine to test if we have indeed or not a
touchscreen device.

Upstream-Status: Submitted
See: https://gitlab.freedesktop.org/wayland/weston/-/merge_requests/1095

Bug-AGL: SPEC-4656
Change-Id: I974b4c5ea8a85f1b3efe4a114f345de948441300
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28314
Tested-by: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>