agl-compositor_git: Bump agl-compositor SRCREV
This SRCREV bump includes the addition of agl-shell-grpc-server, an
interface to agl-shell-desktop protocol, but instead of using wayland
connection it uses Google's RPC. For the time being there's only a
client using it, and that is agl-shell-activator.

The plan is to phase out agl-shell-desktop entirely and use this
interface as a way to perform additional window management.

Further more this bump also includes a couple of connector hot-plug
fixes and renames hide-cursor to disable-cursor entry section to better
reflect its name.

In more detail the following changes have been added:

- 7ef34e0c465a7596 shell: Verify agl-shell version for bound_ok/bound_fail
- c01b90ece22cc587 shell: Add fallback agl-shell version 1
- 0907e87f1576f8ad input: Fix minor issue while reporting disable cursor
- c9d3edb4ee23a563 input: Rename hide-cursor to disable-cursor
- 76fa1b83638310c2 src: Verify if the output is present
- 79eac214d4866d84 src: Further hotplug connector fixes
- 59375972f5642b7e grpc-proxy: Init gRPC
- 0b766cf978b8b100 shell: Not not override the agl_shell interface resources
- c42f50b567a0d5ad shell: Add a wrappers for sending events with agl-shell
- 05bb0384732480a9 compositor: Allow to start other clients
- 14a1292a39377472 agl-shell.xml: Introduce a new interface
- 33997a7c96838795 compositor: Handle output destruction deferral

Change-Id: I6df1d7fcf6d5d3abcc3c1c2b8a0cf820368a26de
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28318
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>