connman-conf: ignore tap interfaces
Add tap* interfaces to the list of interface names that connman
should ignore so that it will not conflict with the tap interfaces
qemu may set up.  This seems safe in general, as a usecase where it
makes sense for connman to assign an address to a tap interface is
not immediately obvious.

Bug-AGL: SPEC-4618

Change-Id: Ic6e82991ccc98496da37b291b3504ed0dc7534c1
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28391
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>