connman-conf: ignore tap interfaces
Add tap* interfaces to the list of interface names that connman
should ignore so that it will not conflict with the tap interfaces
qemu may set up.  This seems safe in general, as a usecase where it
makes sense for connman to assign an address to a tap interface is
not immediately obvious.

Bug-AGL: SPEC-4618

Change-Id: Ic9185e4f93a44c2d9e4265513afc04a2ec4c3d4a
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
(cherry picked from commit 29afa2ad9f9d04477c7a80d3a35e83fb6e5227dc)
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28416
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>