meta-agl-bsp: make sure weston user is in display group for rcar3
Add a meta-rcar-gen3 specific weston-init bbappend to add the
weston user to the display group.  This is preferable to unmasking
the bbappend in upstream meta-rcar-gen3, as we do not want any of
the other changes it contains (or may contain).  This change is a
bit of a stopgap required for running weston or agl-compositor as
the weston user without the agl-driver user session unit in e.g.
the agl-image-weston image or the KVM or IC host images.
Future work will try to align the required groups with upstream
systemd-udevd expectations.

Bug-AGL: SPEC-4618

Change-Id: I22e8aaa095114802242fb1ed150dec2a5055fda9
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
(cherry picked from commit c1b436f4882737ce0b9a95853954d0f4daab2d86)
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28419
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>