meta-agl-bsp: make sure weston user is in display group for rcar3
Add a meta-rcar-gen3 specific weston-init bbappend to add the
weston user to the display group.  This is preferable to unmasking
the bbappend in upstream meta-rcar-gen3, as we do not want any of
the other changes it contains (or may contain).  This change is a
bit of a stopgap required for running weston or agl-compositor as
the weston user without the agl-driver user session unit in e.g.
the agl-image-weston image or the KVM or IC host images.
Future work will try to align the required groups with upstream
systemd-udevd expectations.

Bug-AGL: SPEC-4618

Change-Id: I36bbdc3bfb7ee4de8f77858965693d3c57605482
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28394
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>