meta-core/recipes-graphics/mesa: Backport a fix from 22.2.5
This adds
0001-virgl-don-t-a-use-staging-when-a-resources-created-w.patch which is
a backport from mesa 22.2.5 as a tentative fix for running on virgl on the x86
platform.

This issue seems have been fixed 22.2.4 and 22.2.5, but the original
change was initially quite a long time before in 22.0.X and we seem to
be using 22.0.3.

Bug-AGL: SPEC-4724
Change-Id: I255e30d8e513b8f3bbfa720a24c1afcbd40af2c9
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/28604
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>