meta-agl-core: fix polkit dependencies
As part of the systemd user session removal, some changes were
made to attempt to fix pulling in polkit for core-image-weston,
agl-image-weston, and agl-image-compositor so that the compositors
would correctly start up in the resulting images.  Unfortunately,
just gating the dependency on polkit being in DISTRO_FEATURES breaks
the usecase of using poky and meta-agl-core with our poky-agl distro
configuration without also pulling in the meta-oe layer.  This
combination is used for testing meta-agl-core on the Yocto Project
autobuilder, so that broke and was reported by upstream.

To fix this usecase, the polkit dependency logic has been shifted to
weston-init and agl-compositor-init via bbappends that are only used
when meta-oe is present via our existing BBFILES_DYNAMIC set up.
Since the dependency on polkit stems from a combination of the
compositor PAM configuration from the -init packages and having
systemd, pam, and polkit in DISTRO_FEATURES, this seems more correct,
anyways.

Other changes:
- The explicit pam DISTRO_FEATURES checking that was added in the
  previous cleanup has been removed; it is actually redundant due to
  weston and weston-init having checks.  Using the .inc file for that
  from upstream has been added to the agl-compositor and
  agl-compositor-init recipes to match the weston recipes and make
  the dependency visible in our tree.
- There is no further need for our bbappend to core-image-weston, so
  it has been removed.

Bug-AGL: SPEC-4836

Change-Id: Ia3052023dfb06d1c2190a03c72606ab4b27bbdd2
Signed-off-by: Scott Murray <scott.murray@konsulko.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/29014
Tested-by: Jenkins Job builder account
ci-image-build: Jenkins Job builder account
ci-image-boot-test: Jenkins Job builder account
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>