weston-ini-conf: Add activate-by-default=false by default
As all our platforms (flutter, qt, html5) are now capable of starting
and activating windows on their own, we can turn off the implicit
activation by default from the compositor.

We would remove it entirely from the configuration file once we won't
have any code in the compositor. Make it default for now.

Change-Id: I43df54cd2bba5f05612989b02ce99c63f93a5c2d
Signed-off-by: Marius Vlad <marius.vlad@collabora.com>
Reviewed-on: https://gerrit.automotivelinux.org/gerrit/c/AGL/meta-agl/+/29018
Reviewed-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>
Tested-by: Jan-Simon Moeller <jsmoeller@linuxfoundation.org>